requirements.txt
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg==3.1.19
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
